+++ src/Metadata/class-front-page-builder.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Page For Posts Metadata Builder class
+ * Front Page Metadata Builder class
  *
  * @package Parsely
  * @since 3.4.0
@@ -8,15 +8,15 @@
 
 declare(strict_types=1);
 
-namespace Parsely\Meta;
+namespace Parsely\Metadata;
 
 /**
  * Implements abstract Metadata Builder class to generate the metadata array
- * for a page for posts.
+ * for a front page.
  *
  * @since 3.4.0
  */
-class Page_For_Posts_Builder extends Metadata_Builder {
+class Front_Page_Builder extends Metadata_Builder {
 	/**
 	 * Generates the metadata object by calling the build_* methods and
 	 * returns the value.
@@ -39,6 +39,15 @@ class Page_For_Posts_Builder extends Metadata_Builder {
 	 * @since 3.4.0
 	 */
 	private function build_headline(): void {
-		$this->metadata['headline'] = get_the_title( get_option( 'page_for_posts', true ) );
+		$this->metadata['headline'] = $this->clean_value( get_bloginfo( 'name', 'raw' ) );
+	}
+
+	/**
+	 * Populates the `url` field in the metadata object by getting the home URL.
+	 *
+	 * @since 3.4.0
+	 */
+	protected function build_url(): void {
+		$this->metadata['url'] = home_url();
 	}
 }
